+++ tests/e2e/cases/fs_read_file_buffer_options/in.ts
@@ -15,6 +15,5 @@ console.log("direct:", Buffer.isBuffer(direct), direct.length, direct.readUInt8(
 console.log("object:", Buffer.isBuffer(object), object.length, object.readUInt8(0), object.readUInt8(1), object.readUInt8(2));
 fs.promises.readFile(path, bufferOptions).then((promise: Buffer): void => {
     console.log("promise:", Buffer.isBuffer(promise), promise.length, promise.readUInt8(0), promise.readUInt8(1), promise.readUInt8(2));
+    fs.rmSync(path, { force: true });
 });
-
-fs.rmSync(path, { force: true });